// apps/marketing/marketing-mfe/src/app/http/cafe/product-holding/constants/product-characteristic.enum.ts
export enum ProductCharacteristicEnum {
  PRODUCT_GROUP = 'productGroup',
  PRODUCT_LEVEL = 'productLevel',
  PRODUCT_TIER = 'productTier',
  MAX_MOBILE_LINES = 'maxMobileLines',
  SHARED_MOBILE_DATA = 'sharedMobileData',
}

export enum ProductGroupEnum {
  FMC = 'FMC',
}

export enum ProductLevelEnum {
  LEVEL_2P = '2P',
  LEVEL_3P = '3P',
  LEVEL_4P = '4P',
}

export enum ProductTierEnum {
  LOW = 'Low',
  MID = 'Mid',
  HIGH = 'High',
}